Add signatures mod
